+++ djangocms_lotus/models/article.py
@@ -1,65 +1,82 @@
+from django.conf import settings
 from django.db import models
-from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
-from django.urls import reverse
 
-from .blog import Blog
+from cms.models.pluginmodel import CMSPlugin
+
+from ..choices import (
+    get_latestflux_template_choices,
+    get_latestflux_template_default,
+)
 
 
-class Article(models.Model):
+class ArticleFlux(CMSPlugin):
     """
-    A simple article for a blog.
+    Article flux plugin parameters.
 
     Attributes:
-        blog (models.ForeignKey): Required related Blog object.
-        title (models.CharField): Required title string.
-        content (models.TextField): Optionnal text content.
-        publish_start (models.DateTimeField): Required publication date determine
-            when article will be available.
+        title (models.CharField): An optional title string.
+        template (models.CharField): Template choice from available plugin templates
+            in setting ``CMSLOTUS_ARTICLE_FLUX_TEMPLATES``. Default to the
+            first choice item.
+        length (models.PositiveSmallIntegerField): Required positive small integer.
+        featured_only (models.BooleanField): Optional boolean.
+        from_categories (models.ManyToManyField): Optional choices of Lotus Category
+            objects.
+        from_tags (models.ManyToManyField): Optional choice of Taggit Tag objects.
     """
-    blog = models.ForeignKey(
-        Blog,
-        verbose_name="Related blog",
-        on_delete=models.CASCADE
-    )
 
     title = models.CharField(
         _("title"),
-        blank=False,
+        blank=True,
         max_length=150,
         default="",
     )
-
-    content = models.TextField(
-        _("content"),
-        blank=True,
-        default="",
+    template = models.CharField(
+        _("template"),
+        blank=False,
+        max_length=150,
+        choices=get_latestflux_template_choices(),
+        default=get_latestflux_template_default(),
     )
-
-    publish_start = models.DateTimeField(
-        _("publication start"),
-        db_index=True,
-        default=timezone.now,
+    length = models.PositiveSmallIntegerField(
+        ("length"),
+        default=settings.CMSLOTUS_ARTICLE_FLUX_LIMIT_DEFAULT,
+    )
+    featured_only = models.BooleanField(
+        ("featured only"),
+        default=False,
+        blank=True,
+    )
+    from_categories = models.ManyToManyField(
+        "lotus.Category",
+        verbose_name=_("from categories"),
+        related_name="articleflux",
+        blank=True,
+    )
+    from_tags = models.ManyToManyField(
+        "taggit.Tag",
+        verbose_name=_("from tags"),
+        blank=True,
     )
 
     class Meta:
-        verbose_name = _("Article")
-        verbose_name_plural = _("Articles")
-        ordering = [
-            "-publish_start",
-        ]
+        verbose_name = _("Article flux")
+        verbose_name_plural = _("Article flux")
 
     def __str__(self):
         return self.title
 
-    def get_absolute_url(self):
+    def copy_relations(self, oldinstance):
         """
-        Return absolute URL to the article detail view.
+        Copy relations when plugin object is copied as another object.
 
-        Returns:
-            string: An URL.
+        See:
+
+        https://docs.django-cms.org/en/latest/how_to/09-custom_plugins.html#handling-relations
+
+        NOTE: Not sure we should use this for categories since they are for specific
+        language and plugin can't know about target page language to know if we have
+        to adjust or not depending language.
         """
-        return reverse("djangocms_lotus:article-detail", args=[
-            str(self.blog.id),
-            str(self.id)
-        ])
+        self.from_tags.set(oldinstance.from_tags.all())