+++ CHANGELOG.rst
@@ -6,8 +6,19 @@ Changelog
 Development
 ***********
 
+.. todo::
+    * [ ] Test coverage;
+    * [ ] Include option values in default template;
+    * [ ] Enabled django debug toolbar for development;
+    * [ ] Ensure querysets efficiency helped with debug toolbar;
+    * [ ] Categories choices with language;
+    * [ ] Could we post-save clean category choices depending target language during plugin
+      form save ?
+
 * Adjusted requirements for Lotus;
 * Adjusted Django and Python supports;
+* Added first working version for plugin "Article flux";
+* Included stylesheet patch (for django-admin-style) for Lotus admin category tree view;
 
 
 Version 0.1.0 - Unreleased